Fix super invocations for onnx and tflite backends
The `super()` calls are using the wrong class name. By doing this,
they're skipping the `__init__()` for the actual super class.

This breaks calling `predict()` because `self.lock` has not been
assigned.

The change for the `tflite` backend has been tested. The change for
`onnx` has not been tested, but seems correct.